[#1] Rename: 폴더 구조 변경
